requirements.txt
altair==5.5.0
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
cohere==5.20.0
colorama==0.4.6
cryptography==46.0.3
deprecation==2.1.0
distlib==0.4.0
distro==1.9.0
einops==0.8.1
et_xmlfile==2.0.0
fastavro==1.12.1
filelock==3.20.0
filetype==1.2.0
Flask==3.1.0
Flask-Cors==5.0.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-generativeai
googleapis-common-protos==1.72.0
gotrue==2.12.4
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.36.0
hyperframe==6.1.0
identify==2.6.15
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
json_repair==0.54.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
loguru==0.7.2
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
narwhals==2.12.0
networkx==3.6.1
nodeenv==1.10.0
numpy==2.2.6
oauthlib==3.3.1
openai==1.54.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.2
packaging==25.0
paddleocr==3.3.2
pandas==2.3.3
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.3
pgvector==0.2.4
pillow==10.4.0
platformdirs==4.5.1
playwright==1.57.0
postgrest==0.18.0
pre_commit==4.5.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.8
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pyee==13.0.0
PyJWT==2.10.1
pyparsing==3.2.5
pypdf==3.17.4
pypdfium2==4.30.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.0
python-pptx==0.6.21
pytz==2025.2
PyYAML==6.0.2
realtime==2.24.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.29.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
storage3==0.9.0
streamlit==1.50.0
streamlit-pdf-viewer==0.0.26
supabase==2.10.0
supafunc==0.7.0
surya-ocr==0.17.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.3
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.35.4
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
win32_setctime==1.2.0
xlsxwriter==3.2.9
